fix MAC address matching in lpeg syslog grammar
MAC addresses were being matched against a 7 byte expression, when they
should be treated as 6 bytes. This was causing issues in the kernel
syslog grammar with netfilter, but this was not evident as the MAC field
value is generally 14 bytes in the log output (as it includes trailing
bytes representing ethertype information), so it was normally
successfully matching but not parsing the values into the correct Field
members.