Add SSL support for Elasticsearch output
For some reason I cannot identify, I get an error on the setoption
passthrough saying that the first parameter isn't of type tcp{any}.
I can't really figure out why and I haven't spent a lot of time on it,
but if anyone has any ideas, that'd be great. For now, I commented
them out, but I understand if you do not want that merged.