Update derived_taskgroup_costs scheduled query
The previous version of this query would leave rows with partial data in if a
task group spanned more than one day. For example, if taskGroupId A had entries
on the 1st, 2nd and 3rd of the month, then when calculated on the 1st it would
sum all tasks and costs. When the 1st was >5 days ago, it would add a new row
counting all tasks that happened on the 2nd onwards, meaning that the
mean/median values for costs per task group would drop significantly.

This version uses a merge instead of a flat delete to update existing rows if
present. Duplicates have already been removed.